// src/app/lib/generated/models/ImageGenerationStatus.ts
/* tslint:disable */
/* eslint-disable */
/**
 * AI Babe Public API
 * Lorem ipsum
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const ImageGenerationStatus = {
    Started: 'started',
    Queued: 'queued',
    Running: 'running',
    Finished: 'finished',
    Failed: 'failed',
    MaxConcurencyReached: 'max_concurency_reached',
    ModerationException: 'moderation_exception',
    NotEnoughTokens: 'not_enough_tokens',
    DailyFreeLimitReached: 'daily_free_limit_reached'
} as const;
export type ImageGenerationStatus = typeof ImageGenerationStatus[keyof typeof ImageGenerationStatus];


export function instanceOfImageGenerationStatus(value: any): boolean {
    for (const key in ImageGenerationStatus) {
        if (Object.prototype.hasOwnProperty.call(ImageGenerationStatus, key)) {
            if (ImageGenerationStatus[key as keyof typeof ImageGenerationStatus] === value) {
                return true;
            }
        }
    }
    return false;
}

export function ImageGenerationStatusFromJSON(json: any): ImageGenerationStatus {
    return ImageGenerationStatusFromJSONTyped(json, false);
}

export function ImageGenerationStatusFromJSONTyped(json: any, ignoreDiscriminator: boolean): ImageGenerationStatus {
    return json as ImageGenerationStatus;
}

export function ImageGenerationStatusToJSON(value?: ImageGenerationStatus | null): any {
    return value as any;
}

export function ImageGenerationStatusToJSONTyped(value: any, ignoreDiscriminator: boolean): ImageGenerationStatus {
    return value as ImageGenerationStatus;
}